COMP: External_curl: Fix build issue + Build against patched 7.34.0 release
This commit install the project so that all needed headers are all
in the same folder.

It also change the version of curl to 7.34.0 with an additional patch
adding CURL_STATICLIB to curlbuild.h to fix build issue on windows
and unix system.

Finally, it reverts r22728 which was incorrectly addressing the build
error described in the corresponding commit message.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@22739 3bd1e089-480b-0410-8dfb-8563597acbee