ENH: Updating RSNA Quant tutorial for 2013
Rename to more generic test name so it's not tied to a year.
Use new data sets from Sonia
Add screen shot GUI elements and flags and method, capture screen shots at tutorial steps.
Tweaked all parts to match up to tutorial (first part was already tested
in the Vis tutorial, swapped it out for the Ruler)
Updated CMakeLists.txt so that this commit is independent of the Vis tutorial commit.

Issue #3507

git-svn-id: http://svn.slicer.org/Slicer4/trunk@22741 3bd1e089-480b-0410-8dfb-8563597acbee