COMP: Fix error when building against Qt not having SSL support.
Commit r22782 was done under the assumption that Qt would be built with SSL
support. This commit ensures Qt with/without SSL support is supported.

Note that the build system could be improved to detect with Qt has
OpenSSL support, if this is the case, the OpenSSL project could be
automatically enabled unless explicitly disabled.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@22799 3bd1e089-480b-0410-8dfb-8563597acbee