ENH: Add 'pydicom' python module as external project.
"pydicom is a pure python package for working with DICOM files. It
was made for inspecting and modifying DICOM data in an easy "pythonic"
way. The modifications can be written again to a new file. As a pure
python package, it should run anywhere python runs without any other
requirements.

pydicom is not a DICOM server, and is not primarily about viewing
images. It is designed to let you manipulate data elements in DICOM
files with python code.

Limitations -- the main limitation of the current version is that
compressed pixel data (e.g. JPEG) cannot be altered in an intelligent
way as it can for uncompressed pixels. Files can always be read and
saved, but compressed pixel data cannot easily be modified."
See http://code.google.com/p/pydicom/

The module being licensed under MIT license, it is compatible with
the Slicer BSD like license.

Co-authored-by: Andriy Fedorov <fedorov@bwh.harvard.edu>

From: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>

git-svn-id: http://svn.slicer.org/Slicer4/trunk@22815 3bd1e089-480b-0410-8dfb-8563597acbee