BUG: Fix/rewrite setupSlicerEnvironment.sh
Rewrite setupSlicerEnvironment.sh to use the new launcher option to
generate commands to set the environment. This fixes several problems
with the old setupSlicerEnvironment.sh, including:

- Failure to properly quote in order to cope with special characters.
- Incorrectly treating everything as a path list.

The new script is a glorified wrapper around 'eval' and exists more for
backwards compatibility than anything (and also now states that it is
deprecated).

git-svn-id: http://svn.slicer.org/Slicer4/trunk@23038 3bd1e089-480b-0410-8dfb-8563597acbee