BUG: ResampleDTIlog Euclidean was using the measurement frame to transform tensors from local coordinates to world coordinates
It should be using the space direction of the image instead. The measurement frame should only transform gradient vectors from DWI to world coordinates. To be backward compatible, if a measurement frame is detected in the MetaData of the image, the measurement frame is still used as the transform matrix from local to world coordinate unless --noMeasurementFrame is specified. If no measurement frame information is available in the image MetaData, the image space direction is used instead (as well as if --noMeasurementFrame is specified).

From: fbudin69500 <fbudin@unc.edu>

git-svn-id: http://svn.slicer.org/Slicer4/trunk@23103 3bd1e089-480b-0410-8dfb-8563597acbee