BUG: fix behaviour of gui when markup inside transform
When editing the coordinates of a markup that's inside of a
transform, the pairing of getting the untransformed
coordinate and then saving it as transformed was leading to
incorrect values.
Added in a checkbox to select if the GUI table shows the
markup coordinates in transformed or untransformed versions.

Issue #3632

git-svn-id: http://svn.slicer.org/Slicer4/trunk@23114 3bd1e089-480b-0410-8dfb-8563597acbee