COMP: Avoid unnecessary recompilation on MacOSX. Fixes #2583
This has been discussed with the CMake folks. A small example [1]
allowing to reproduce the problem has been provided and the solution is
to use the option "CMAKE_BUILD_WITH_INSTALL_RPATH" to build the
python project.

Note that this fix is not required for zlib and curl project because the
associated library is built statically and is not updated at install
time.

8<----8<----8<----8<----8<----8<----8<----8<----8<----
From Brad King - April 24, 2014:
It's due to the install_name_tool used to update the binary
on installation. The cmake_install.cmake file contains:

    execute_process(COMMAND "/usr/bin/install_name_tool"
      -id "libfoo.dylib"
      "$ENV{DESTDIR}${CMAKE_INSTALL_PREFIX}/lib/libfoo.dylib")

If you build with -DCMAKE_BUILD_WITH_INSTALL_RPATH=ON then
no install-time update will occur.
8<----8<----8<----8<----8<----8<----8<----8<----8<----

[1] https://gist.github.com/jcfr/11233992

git-svn-id: http://svn.slicer.org/Slicer4/trunk@23150 3bd1e089-480b-0410-8dfb-8563597acbee