BUG: #3694 check for valid image data before running pipeline
Since the display node is observing the image data directly it
may receive Modified events on image data that is not yet
completely constructed, in particular, during a DeepCopy, the scalar
fields may not yet have been copied when a ModifiedEvent is triggered
by one of the Set methods.

This change prevents a spurious vtkErrorMessage from being generated,
and so it allows the EMSegment test to pass.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@23223 3bd1e089-480b-0410-8dfb-8563597acbee