BUG: Fix TemplateManager unicode support
Add a helper function (in SlicerWizard.Utilities) to detect the
character encoding of an input byte sequence. Use this in
TemplateManager to correctly handle encoding conversion of template
files in case the replacement text is a unicode, rather than a str
(which otherwise results in an exception when automatic conversion tries
to convert non-ASCII files).

git-svn-id: http://svn.slicer.org/Slicer4/trunk@23266 3bd1e089-480b-0410-8dfb-8563597acbee