BUG: Fix (another) bug reading settings list
Write a helper function to read a QSettings value explicitly as a list
in Python, in order to work around PythonQt's implicit conversion of
QVariant. Use this everywhere it is applicable in the Extension Wizard
module. This allows us to simplify some places where we iterate over
settings lists, and (more importantly) fixes (another) bug where we try
to iterate over a list that might be empty.

See also r23285.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@23302 3bd1e089-480b-0410-8dfb-8563597acbee