ENH: Implement installation of updates
Implement installation of updates to extensions (consisting of first
removing the old version, then installing the new version). Note that
this will attempt to install new dependencies as well. Also, change
handling of update downloads to additionally write the metadata for the
update (in addition to the archive), as we'll want it when we install
the update and need to persist it across a restart.

Issue #2089

git-svn-id: http://svn.slicer.org/Slicer4/trunk@23504 3bd1e089-480b-0410-8dfb-8563597acbee