ENH: Added SceneViews and Segment plugins to subject hierarchy
SceneViews is a role plugin for scene view nodes. Provides an icon and allows restoring scene views (request of Sonia)
Segment is a function plugin that adds a context sub-menu ("Segment this using...") to volumes. There is only one action under the sub-menu so far that takes the user to the Editor module, sets the volume as master and creates a labelmap as merge volume.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@23551 3bd1e089-480b-0410-8dfb-8563597acbee